eliminate most of warnings

+++ source/utils/tiny_obj_loader.h
@@ -344,9 +344,9 @@ struct skin_weight_t {
 // Index struct to support different indices for vtx/normal/texcoord.
 // -1 means not used.
 struct index_t {
-  int vertex_index;
-  int normal_index;
-  int texcoord_index;
+  uint64_t vertex_index;
+  uint64_t normal_index;
+  uint64_t texcoord_index;
 };
 
 struct render_object_t {
